# Tests: cache some binary rocks to speed up tests