The Linux job for "lua=5.1" on Travis went to about 159s to 142s.
The Windows job for "lua=5.1,msvc" on Appveyor went from about 654s to 569s.

Since Appveyor does not run jobs in parallel, the gains are considerable. Total Appveyor time went from 1h40 to 2h20 once all mock-server tests were enabled; this at least brings the total time down to the previous ballpark.