@hishamhm 2018-05-28
@georgeroman I made this into a PR first instead of merging directly so you could take a look!